azure_functions/requirements.txt
-r ../requirements.txt
azure-functions
